refactor: migrate tests to typescript
